loaded data using email and customization preference both using fetch

+++ src/pages/MyArtAndCraftList.jsx
@@ -1,10 +1,10 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../authProvider/AuthProvider";
 import MyItemsCard from "../components/MyItemsCard";
+import Swal from "sweetalert2";
 
 const MyArtAndCraftList = () => {
   const [myItems, setMyItems] = useState([]);
-  const [filteredItems, setFilteredItems] = useState(myItems);
   const [isLoading, setIsLoading] = useState(true);
   const { user } = useContext(AuthContext);
   const email = user.email;
@@ -28,25 +28,43 @@ const MyArtAndCraftList = () => {
 
   // Delete item function
   const handleDelete = (id) => {
-    fetch(`https://eco-craft-server.vercel.app/delete/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => console.log(data));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`https://eco-craft-server.vercel.app/delete/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              const itemAfterDelete = myItems.filter((item) => item._id !== id);
+              console.log(itemAfterDelete);
+              setMyItems(itemAfterDelete);
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+            }
+          });
+      }
+    });
   };
 
   // Customization Filter
   const handleCustomization = (e) => {
     const selectedOption = e.target.value;
 
-    if (selectedOption === "Yes") {
-      const filteredNewItems = myItems.filter((item) => item.customization === "Yes");
-      setFilteredItems(filteredNewItems);
-    }
-    if (selectedOption === "No") {
-      const filteredNewItems = myItems.filter((item) => item.customization === "No");
-      setFilteredItems(filteredNewItems);
-    }
+    fetch(`http://localhost:5000/customization/${email}/${selectedOption}`)
+      .then((res) => res.json())
+      .then((data) => setMyItems(data));
   };
 
   return (
@@ -64,15 +82,15 @@ const MyArtAndCraftList = () => {
           <option value="No">No</option>
         </select>
       </div>
+
       {myItems.length === 0 && (
         <h1 className="text-[32px] flex items-center justify-center font-semibold text-center h-[50vh]">You have not added any item yet</h1>
       )}
-      <div className="grid lg:grid-cols-3 md:grid-cols-2 gap-5">
-        {filteredItems.length > 0 &&
-          filteredItems.map((myItem, index) => <MyItemsCard myItem={myItem} handleDelete={handleDelete} key={index}></MyItemsCard>)}
 
-        {filteredItems.length === 0 &&
-          myItems.map((myItem, index) => <MyItemsCard myItem={myItem} handleDelete={handleDelete} key={index}></MyItemsCard>)}
+      <div className="grid lg:grid-cols-3 md:grid-cols-2 gap-5">
+        {myItems.map((myItem, index) => (
+          <MyItemsCard myItem={myItem} handleDelete={handleDelete} key={index}></MyItemsCard>
+        ))}
       </div>
     </div>
   );
